TwitchEventSub reimplementation
Reference commit 06e7601dd36101f6fdad0b2216e284e7b83c5835

* Updated ConnectAsync to validate user type and start tracking users subscribed.

The reason for this change is because TwitchEventSub requires applications to create subscriptions within 10 seconds of receiving the welcome message. Additionally, clients should not try to subscribe immediately after establishing a connection because the session id is not available until the welcome message is received.

* Updated SubscribeAsync to register notification handler and create user subscriptions if event sub is connected.

* Updated UnsubscribeAsync to delete user subscriptions and remove notification handlers.